% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_packages_long.R
\name{available_packages_long}
\alias{available_packages_long}
\title{Available Packages Long Format}
\usage{
available_packages_long(cran_repo_url = "win_binary_default")
}
\arguments{
\item{cran_repo_url}{The url of the CRAN repository to check.}
}
\value{
Tibble of packages restructured in long format with one row per
imports or depends package.
}
\description{
This creates a tidy format tibble of package dependencies (depends, imports).
}
\examples{
\dontrun{
available_packages_long()
}
}
