% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_packages_tb.R
\name{available_packages_tb}
\alias{available_packages_tb}
\title{Get all available R packages as a tibble.}
\usage{
available_packages_tb(cran_repo_url = "win_binary_default")
}
\arguments{
\item{cran_repo_url}{The url of the CRAN repository to check.}
}
\value{
Tibble of all available packages and details.
}
\description{
The matrix available.packages is available with utils.
This will extract the matrix from the specified CRAN repository
and convert it to a tibble.
}
\examples{
available_packages_tb()
}
