% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r_version.R
\name{check_r_version}
\alias{check_r_version}
\title{Check installed R version meets minimum for packages.}
\usage{
check_r_version(required_packages)
}
\arguments{
\item{required_packages}{Tibble of required packages.}
}
\value{
Tibble of required packages that require R > existing installed
version. Or FALSE if no required packages exceed existing R install version.
}
\description{
Check installed R version meets minimum for packages.
}
\examples{
available_long_tb <- available_packages_long()
required_tb <- search_required_packages(available_long_tb, "fabletools")
exceed_r_version <- check_r_version(required_tb)
}
