% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_urls.R
\name{check_urls}
\alias{check_urls}
\title{Check if urls in a specified column are genuine.}
\usage{
check_urls(check_tb, url_column = "package_url", verify_url_proxy = NULL)
}
\arguments{
\item{check_tb}{Tibble or dataframe with column of urls to check.}

\item{url_column}{Name of column containing urls to check.}

\item{verify_url_proxy}{Optionally specify a proxy when on secure network.}
}
\value{
Input tibble or dataframe with url_ok column added.
}
\description{
A column will be added to the input dataframe "url_ok" with TRUE or FALSE
depending on if URL works.
}
\examples{
search_tb <- search_requirements(
  packages_long = available_packages_long(), package_name = "dplyr"
)
search_tb <- verify_urls(search_tb)
}
