% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_available_server.R
\name{compare_available_server}
\alias{compare_available_server}
\title{Compare tibbles of required packages and existing packages}
\usage{
compare_available_server(required_packages, server_packages)
}
\arguments{
\item{required_packages}{Tibble of required packages.
generated by get_required_packages()}

\item{server_packages}{Tibble of server packages.
generated by existing_server_packages()}
}
\value{
Tibble with status of each required package.
}
\description{
Compare the package and name package versions of the two tibbles. Add a
new column "server_status", specifying if package exists and needs upgrading,
does not exist, so needs to be acquired, or exists at the right version.
}
\examples{
available_long_tb <- available_packages_long()
required_tb <- search_requirements(available_long_tb, "fabletools")
server_tb <- existing_server_packages("//s1428a/R_Packages/R_3_6_3_Packages")
compare_tb <- compare_available_server(required_tb, server_tb)
}
