% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_required_packages.R
\name{search_required_packages}
\alias{search_required_packages}
\title{Search for required packages by navigating tree of package dependencies.}
\usage{
search_required_packages(packages_long, package_name)
}
\arguments{
\item{packages_long}{Tibble of all available packages with dependencies in
long format, i.e. output of available_packages_long()}

\item{package_name}{Name of a specific R package to check}
}
\value{
Tibble showing input package and dependent packages, their minimum R
version and minimum package version.
}
\description{
Packages have requirements on other packages via depends, imports.
}
\examples{
# Get required packages tibble for dplyr
search_required_packages(
  packages_long = available_packages_long(), package_name = "dplyr")

}
