% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_requirements.R
\name{search_requirements}
\alias{search_requirements}
\title{Search for required packages by navigating tree of package dependencies.}
\usage{
search_requirements(packages_long, package_name, package_version_number = NA)
}
\arguments{
\item{packages_long}{Tibble of all available packages with dependencies in
long format, i.e. output of available_packages_long().}

\item{package_name}{Package name to search.}

\item{package_version_number}{Optional version number required. Will show a
message if available version less than or greater than this.}
}
\value{
Tibble showing input package and dependent packages, their minimum R
version and minimum package version.
}
\description{
Packages have requirements on other packages via depends, imports.
}
\examples{
# Get required packages tibble for dplyr
search_requirements(
  packages_long = available_packages_long(), package_name = "dplyr"
)
}
