% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_requirements_multiple.R
\name{search_requirements_multiple}
\alias{search_requirements_multiple}
\title{Find all dependencies for multiple input search packages.}
\usage{
search_requirements_multiple(packages_long, package_names)
}
\arguments{
\item{packages_long}{Tibble of all available packages with dependencies in
long format, i.e. output of available_packages_long().}

\item{package_names}{Vector of package names to search.}
}
\value{
Tibble showing input packages and dependent packages, their minimum R
version and minimum package version.
}
\description{
Find all dependencies for multiple input search packages.
}
\examples{
#' \dontrun{
# Search for all existing packages on a server...
# Get existing server packages
server_tb <- existing_server_packages("my_server/R_Packages/R_3_6_3_Packages")
# Search for their requirements by passing the server_package column vector
search_requirements_multiple(
  packages_long = available_packages_long(),
  package_names = server_tb$server_package
)
}
}
